IsoJump, Hexagon Library, Water SSR PostProcess tests.
Moved hexagon math into a static library script accessible with HexLib.func()
IsoJump2D is a bridge (C) between the Ysort root Node (B) (inherited Y Sort only, from A). This is how jumping works as expected with YSortEnabled in Godot! Lots of intermediaries, but in the future we can likely settle on a common structure for characters going forwards that does all this with minimal issue. Including ensuring all characters exist in the same YSorted "world".