Rivers, Spawn Anything, and hex format logics
Looks like the best way to do rivers will be with the built in Line2D tool - as it makes a nice textured 2D mesh we can use to plop down where we want a river water. Will need to work out how to give it the special water treatment in the future when feeling Shady.
Pulled the HexTile code controlling the spawn effect out into its own dedicated Node. It can now just be attached to anything you want to spawn in dramatically. Hurrah.
Satrted on the hexagon format logic, which will be used to spawn in the correct Hexagon Format in the future for rivers and cliffs.